+++ src/styles/emotion.d.ts
@@ -2,5 +2,6 @@ import '@emotion/react';
 import { ThemeType } from './theme';
 
 declare module '@emotion/react' {
+  // eslint-disable-next-line @typescript-eslint/no-empty-object-type
   export interface Theme extends ThemeType {}
 }
